// src/modules/users/interfaces.ts
export enum SocialProviderEnum {
  GOOGLE = 'google',
  FACEBOOK = 'facebook',
  TWITTER = 'twitter',
  NONE = 'none',
}

export enum BodyHairEnum {
  HAIRY = 'hairy',
  TRIMMED = 'trimmed',
  SMOOTH = 'smooth',
  WAXED = 'waxed',
  SHAVED = 'shaved',
  PARTIALLY_HAIRY = 'partially hairy',
}

export enum EyeColorEnum {
  HONEY = 'honey',
  LIGHT_BROWN = 'light brown',
  DARK_BROWN = 'dark brown',
  LIGHT_BLUE = 'light blue',
  DARK_BLUE = 'dark blue',
  LIGHT_GREEN = 'light green',
  DARK_GREEN = 'dark green',
  BLACK = 'black',
  GRAY = 'gray',
  HAZEL = 'hazel',
  AMBER = 'amber',
  HETEROCROMIA = 'heterochromia',
  OTHER = 'other',
}
export enum HairColorEnum {
  BLONDE = 'blonde',
  BRUNETTE = 'brunette',
  LIGHT_BROWN = 'light brown',
  DARK_BROWN = 'dark brown',
  BLACK = 'black',
  RED = 'red',
  AUBURN = 'auburn',
  GRAY = 'gray',
  WHITE = 'white',
  BALD = 'bald',
  DYED = 'dyed',
  OTHER = 'other',
}

export enum SkinColorEnum {
  TYPE_1 = 'type_1', // Pele muito clara, queima facilmente, nunca bronzeia
  TYPE_2 = 'type_2', // Pele clara, queima com facilidade, bronzeia pouco
  TYPE_3 = 'type_3', // Pele morena clara, queima moderadamente, bronzeia gradualmente
  TYPE_4 = 'type_4', // Pele morena moderada, queima pouco, bronzeia bem
  TYPE_5 = 'type_5', // Pele morena escura, raramente queima, bronzeia muito
  TYPE_6 = 'type_6', // Pele muito escura, nunca queima, altamente pigmentada
  MIXED = 'mixed', // Para pessoas com múltiplas etnias/tonalidades
  OTHER = 'other', // Caso alguém prefira não se enquadrar
}

export enum TribeEnum {
  TWINK = 'twink',
  BEAR = 'bear',
  CHUBBY = 'chubby',
  CHASER = 'cahser',
  DADDY = 'daddy',
  JOCK = 'jock',
  DRAG = 'drag',
  OTTER = 'otter',
  GEEK = 'geek',
  LEATHER = 'leather',
  CUB = 'cub',
  CHUB = 'chub',
  TWUNK = 'twunk',
  WOLF = 'wolf',
  BUTCH = 'butch',
  FEMME = 'femme',
  ANDROGYNOUS = 'androgynous',
  TRANS = 'trans',
  QUEER = 'queer',
  NON_BINARY = 'non binary',
  GENDERFLUID = 'genderfluid',
  OTHER = 'other',
}

export enum RelationshipStatusEnum {
  SINGLE = 'single',
  MARRIED = 'married',
  OPEN_RELATIONSHIP = 'open_relationship',
  WIDOWED = 'widowed',
  DIVORCED = 'divorced',
  SEPARATED = 'separated',
  COMPLICATED = 'complicated',
  POLYAMOROUS = 'polyamorous',
}

export enum SexualPreferenceEnum {
  TOP = 'top',
  BOTTOM = 'bottom',
  VERSATILE = 'versatile',
  VERSATILE_TOP = 'versatile + top',
  VERSATILE_BOTTOM = 'versatile + bottom',
  NO_PENETRATION = 'no penetration',
}